folder/requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
fasttext==0.9.3
func-argparse==1.1.1
idna==3.11
joblib==1.5.3
numpy==2.2.6
pandas==2.3.3
psutil==7.2.1
pybind11==3.0.1
pybloomfiltermmap3==0.6.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
requests==2.32.5
sacremoses==0.1.1
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
soupsieve==2.8.1
submitit==1.5.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
uniseg==0.10.0
urllib3==2.6.2
